refactor: User -> Member로 변경
